+++ frontend/src/lib/internalMetrics.ts
@@ -1,5 +1,6 @@
 import posthog from 'posthog-js'
-import api from 'lib/api'
+import api, { getJSONOrThrow } from 'lib/api'
+import { getResponseBytes } from 'scenes/insights/utils'
 
 export interface TimeToSeeDataPayload {
     type: 'insight_load' | 'dashboard_load'
@@ -33,4 +34,37 @@ export async function captureTimeToSeeData(teamId: number | null, payload: TimeT
             ...payload,
         })
     }
+}
+
+/** api.get() wrapped in captureTimeToSeeData() tracking for simple cases of fetching insights or dashboards.
+ * This is not in api.ts to avoid circular dependencies, but the principle is the same.
+ */
+export async function apiGetWithTimeToSeeDataTracking<T extends any>(
+    url: string,
+    teamId: number | null,
+    timeToSeeDataPayload: Omit<
+        TimeToSeeDataPayload,
+        'api_url' | 'time_to_see_data_ms' | 'status' | 'api_response_bytes'
+    >
+): Promise<T> {
+    let response: Response | undefined
+    let error: any
+    const requestStartMs = performance.now()
+    try {
+        response = await api.getResponse(url)
+    } catch (e) {
+        error = e
+    }
+    const requestDurationMs = performance.now() - requestStartMs
+    captureTimeToSeeData(teamId, {
+        ...timeToSeeDataPayload,
+        api_url: url,
+        status: error ? 'failure' : 'success',
+        api_response_bytes: response && getResponseBytes(response),
+        time_to_see_data_ms: requestDurationMs,
+    })
+    if (!response) {
+        throw error // If there's no response, there must have been an error - rethrow it
+    }
+    return (await getJSONOrThrow(response)) as T
 }
